requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.2
flake8==6.0.0
inflection==0.5.1
isort==5.12.0
jsonschema==4.17.3
mccabe==0.7.0
Pillow==9.5.0
psycopg2==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
